+++ Sources/ConnectSDK/Recovery/ConnectRecoverySession.swift
@@ -18,19 +18,17 @@ public class ConnectRecoverySession {
     private let theme: Theme
     private let callbacks: RecoveryCallbacks
     private let allowList: ConnectAllowList
-    private let oauthCallback: ConnectOAuthCallback
     private var isPresented: Bool = false
     private var activeSession: ConnectSession?
 
     // MARK: - Initialization
 
-    internal init(jwt: String, environment: Environment, theme: Theme, callbacks: RecoveryCallbacks, allowList: ConnectAllowList = .default, oauthCallback: ConnectOAuthCallback = .default) {
+    internal init(jwt: String, environment: Environment, theme: Theme, callbacks: RecoveryCallbacks, allowList: ConnectAllowList = .default) {
         self.jwt = jwt
         self.environment = environment
         self.theme = theme
         self.callbacks = callbacks
         self.allowList = allowList
-        self.oauthCallback = oauthCallback
     }
 
     // MARK: - Public Methods
@@ -58,7 +56,6 @@ public class ConnectRecoverySession {
             theme: theme.rawValue,
             callbackHandler: callbackHandler,
             allowList: allowList,
-            oauthCallback: oauthCallback
         )
 
         let nav = UINavigationController(rootViewController: webVC)